+++ kernel/core/datamanagement/cluster.cpp
@@ -90,7 +90,7 @@ namespace NumeRe
             nGlobalType = ClusterItem::ITEMTYPE_INVALID;
 
         if (nNum == 1)
-            _idx.row.setOpenEndIndex(std::max(_idx.row.front(), (int)size()) - 1);
+            _idx.row.setOpenEndIndex(std::max((size_t)_idx.row.front(), size()) - 1);
 
         // Assign the single results
         for (size_t i = 0; i < _idx.row.size(); i++)
